New assembly for native interop was added to the solution. XML docs were added.
