+++ src/Dali/Dali.sln
@@ -7,9 +7,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedSharp.Dali", "Dali\RedSh
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedSharp.Dali.ViewModel", "RedSharp.Dali.ViewModel\RedSharp.Dali.ViewModel.csproj", "{2F5CA780-944A-4918-A8FD-A4CE6DE99D98}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RedSharp.Dali.Common", "RedSharp.Dali.Common\RedSharp.Dali.Common.csproj", "{5D8095F5-61C8-4117-959F-5F9224B1F7F5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedSharp.Dali.Common", "RedSharp.Dali.Common\RedSharp.Dali.Common.csproj", "{5D8095F5-61C8-4117-959F-5F9224B1F7F5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RedSharp.Dali.Controls", "RedSharp.Dali.Controls\RedSharp.Dali.Controls.csproj", "{473FCCBB-597C-4CD6-AE33-1F3226AAFE5B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedSharp.Dali.Controls", "RedSharp.Dali.Controls\RedSharp.Dali.Controls.csproj", "{473FCCBB-597C-4CD6-AE33-1F3226AAFE5B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RedSharp.Dali.Common.Interop", "RedSharp.Dali.Common.Interop\RedSharp.Dali.Common.Interop.csproj", "{56E0E389-243E-45CD-822B-AA43AC1933FA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -69,6 +71,18 @@ Global
 		{473FCCBB-597C-4CD6-AE33-1F3226AAFE5B}.Release|x64.Build.0 = Release|Any CPU
 		{473FCCBB-597C-4CD6-AE33-1F3226AAFE5B}.Release|x86.ActiveCfg = Release|Any CPU
 		{473FCCBB-597C-4CD6-AE33-1F3226AAFE5B}.Release|x86.Build.0 = Release|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Debug|x64.Build.0 = Debug|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Debug|x86.Build.0 = Debug|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Release|x64.ActiveCfg = Release|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Release|x64.Build.0 = Release|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Release|x86.ActiveCfg = Release|Any CPU
+		{56E0E389-243E-45CD-822B-AA43AC1933FA}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
