+++ src/Dali/Dali/View/MainWindow.xaml
@@ -7,6 +7,6 @@
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
     <Grid>
-        <Button Width="120" Height="30" Click="Button_Click">Click</Button>
+        
     </Grid>
 </Window>
