+++ src/Dali/RedSharp.Dali.Common.Interop/RedSharp.Dali.Common.Interop.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <UseWPF>true</UseWPF>
+    <UseWindowsForms>true</UseWindowsForms>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\RedSharp.Dali.Common\RedSharp.Dali.Common.csproj" />
+  </ItemGroup>
+
+</Project>
