+++ src/Dali/RedSharp.Dali.Common/Enums/HotkeyModifier.cs
@@ -2,11 +2,14 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace RedSharp.Dali.Common.GlobalHotkey
+namespace RedSharp.Dali.Common.Enums
 {
     /// <summary>
-    /// 
+    /// List of modify buttons for register global HotKeys.
     /// </summary>
+    /// <SecurityNote>
+    /// Do not change values.
+    /// </SecurityNote>
     public enum HotkeyModifier
     {
         None = 0,
