+++ src/Dali/RedSharp.Dali.Common/RedSharp.Dali.Common.csproj
@@ -1,8 +1,7 @@
-<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
-
-    <PropertyGroup>
+<Project Sdk="Microsoft.NET.Sdk">
+  
+  <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-    <UseWPF>true</UseWPF>
   </PropertyGroup>
 
 </Project>
